feat: Add Clone Magnet Button
Adds a "Clone Magnet" button to the Lead Magnet Card component, allowing users to clone a lead magnet and customize it.

The button triggers a dialog that prompts the user to confirm their action. If the user is logged in, they can clone the magnet and be redirected to the builder. If they are not logged in, they are redirected to the login page.